+++ assembly/memory/memory-map.ts
@@ -3,7 +3,7 @@ import MemoryAccessor from './memory-accessor';
 export default interface MemoryMap {
     read32(address: u32, accessor: MemoryAccessor): u32;
 
-    read16(address: u32, accessor: MemoryAccessor): u32;
+    read16(address: u32, accessor: MemoryAccessor): u16;
 
     read8(address: u32, accessor: MemoryAccessor): u8;
